refactor: split textual.cc into 3 files along responsibility boundaries
Split the 2,132-line textual.cc parser into focused modules:
- textual_internal.h (206 lines): shared instance_t class declaration,
  application_t struct, and parse_amount_expr in ledger::detail namespace
- textual.cc (286 lines): core parse loop, line reading, entry points
- textual_directives.cc (787 lines): all directive handlers and
  general_directive dispatch
- textual_xacts.cc (979 lines): transaction and posting parsing

The instance_t class moves from an anonymous namespace to
ledger::detail to enable sharing across translation units. All 1,392
tests pass with no behavioral changes.
